+++ koe/models.py
@@ -357,8 +357,8 @@ class Segment(SimpleModel):
     # Time ID - unique for each combination of (song name, start and end). Recalculate if end/begin changes
     tid = models.IntegerField(null=True, blank=False)
 
-    start_time_ms = models.IntegerField()
-    end_time_ms = models.IntegerField()
+    start_time_ms = models.FloatField()
+    end_time_ms = models.FloatField()
 
     audio_file = models.ForeignKey(AudioFile, on_delete=models.CASCADE)
